UniversalKafkaDeserializerFactory and SchemaRegistryBasedDeserializerFactory flatten
