+++ utils/schemed-kafka-components-utils/src/main/scala/pl/touk/nussknacker/engine/schemedkafka/schemaregistry/serialization/KafkaSchemaRegistryBasedKeyValueDeserializationSchemaFactory.scala
@@ -1,0 +1,37 @@
+package pl.touk.nussknacker.engine.schemedkafka.schemaregistry.serialization
+
+import io.confluent.kafka.schemaregistry.ParsedSchema
+import org.apache.kafka.common.serialization.Deserializer
+import pl.touk.nussknacker.engine.kafka.KafkaConfig
+import pl.touk.nussknacker.engine.schemedkafka.RuntimeSchemaData
+import pl.touk.nussknacker.engine.schemedkafka.schemaregistry.SchemaRegistryClientFactory
+import pl.touk.nussknacker.engine.schemedkafka.schemaregistry.universal.UniversalKafkaDeserializerFactory
+import pl.touk.nussknacker.engine.schemedkafka.serialization.KafkaSchemaBasedKeyValueDeserializationSchemaFactory
+
+class KafkaSchemaRegistryBasedKeyValueDeserializationSchemaFactory(
+    protected val schemaRegistryClientFactory: SchemaRegistryClientFactory,
+    protected val deserializerFactory: UniversalKafkaDeserializerFactory
+) extends KafkaSchemaBasedKeyValueDeserializationSchemaFactory {
+
+  override protected def createKeyDeserializer[K](
+      schemaDataOpt: Option[RuntimeSchemaData[ParsedSchema]],
+      kafkaConfig: KafkaConfig
+  ): Deserializer[K] =
+    createDeserializer[K](kafkaConfig, schemaDataOpt, isKey = true)
+
+  override protected def createValueDeserializer[V](
+      schemaDataOpt: Option[RuntimeSchemaData[ParsedSchema]],
+      kafkaConfig: KafkaConfig
+  ): Deserializer[V] =
+    createDeserializer[V](kafkaConfig, schemaDataOpt, isKey = false)
+
+  private final def createDeserializer[T](
+      kafkaConfig: KafkaConfig,
+      schemaDataOpt: Option[RuntimeSchemaData[ParsedSchema]],
+      isKey: Boolean
+  ): Deserializer[T] = {
+    val schemaRegistryClient = schemaRegistryClientFactory.create(kafkaConfig)
+    deserializerFactory.createDeserializer[T](schemaRegistryClient, kafkaConfig, schemaDataOpt, isKey)
+  }
+
+}
